+++ go.mod
@@ -1,5 +1,8 @@
-module github.com/elliotforbes/go-websocket-tutorial
+module proj6/gomoon
 
 go 1.18
 
-require github.com/gorilla/websocket v1.5.0
+require (
+	github.com/go-chi/chi/v5 v5.0.7
+	github.com/gorilla/websocket v1.5.0
+)
