+++ routes/auth
@@ -1,0 +1,21 @@
+package routes
+
+import (
+
+	"github.com/markbates/goth/gothic"
+
+	"net/http"
+
+)
+
+func handleRouteAuth(http_ http) {
+
+	http_.HandleFunc("auth/login", func(res http.ResponseWriter, req *http.Request) {
+		// try to get the user without re-authenticating
+		if gothUser, err := gothic.CompleteUserAuth(res, req); err == nil {
+		} else {
+			gothic.BeginAuthHandler(res, req)
+		}
+	})
+
+}
