+++ server.go
@@ -1,106 +1,27 @@
 package main
 
 import (
-	"fmt"
-	"log"
 	"net/http"
 
-	"github.com/gorilla/websocket"
+	"github.com/go-chi/chi/v5"
+    "log"
+	"proj6/gomoon/routes"
 )
 
-var users = []*websocket.Conn{}
 
-var upgrader = websocket.Upgrader{
-	ReadBufferSize:  1024,
-	WriteBufferSize: 1024,
-}
-
-func helloHandler(w http.ResponseWriter, r *http.Request) {
-	if r.URL.Path != "/hello" {
-		http.Error(w, "404 not found, from handler.", http.StatusNotFound)
-		return
-	}
-
-	if r.Method != "GET" {
-		http.Error(w, "Method is not supported, from handler.", http.StatusNotFound)
-		return
-	}
-
-	fmt.Fprintf(w, "Hello!")
-}
-
-func reader(conn *websocket.Conn) {
-	for {
-		// read in a message
-		messageType, p, err := conn.ReadMessage()
-		if err != nil {
-			log.Println(err)
-			return
-		}
-
-		// print out that message for clarity
-		log.Println("sending from here")
-		fmt.Println("text" + string(p))
-
-		log.Println(users)
-
-		Boardcast(users, p)
-
-		if err := conn.WriteMessage(messageType, p); err != nil {
-			log.Println(err)
-			return
-		}
-	}
-}
-
-func wsEndpoint(w http.ResponseWriter, r *http.Request) {
-	upgrader.CheckOrigin = func(r *http.Request) bool { return true }
-
-	// upgrade this connection to a WebSocket
-	// connection
-	ws, err := upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		log.Println(err)
-	}
-
-	log.Println("Client Connected")
-	err = ws.WriteMessage(1, []byte("Hi Client!"))
-	if err != nil {
-		log.Println(err)
-	}
-	// listen indefinitely for new messages coming
-	// through on our WebSocket connection
-	users = append(users, ws)
-
-	reader(ws)
-
-}
-
-func formHandler(w http.ResponseWriter, r *http.Request) {
-	if err := r.ParseForm(); err != nil {
-		fmt.Fprintf(w, "ParseForm() err: %v", err)
-		return
-	}
-	fmt.Fprintf(w, "POST from post handler "+r.Method)
-	name := r.FormValue("name")
-	address := r.FormValue("address")
-
-	fmt.Fprintf(w, "Name = %s\n", name)
-	fmt.Fprintf(w, "Address = %s\n", address)
-}
 
 func main() {
-	fileServer := http.FileServer(http.Dir("./static"))
 
-	http.HandleFunc("/form", formHandler)
-	http.HandleFunc("/hello", helloHandler)
-	http.Handle("/", fileServer)
+	r := chi.NewRouter()
 
-	fmt.Printf("Starting server at port 8080\n")
+	r.Mount("/", routes.StaticRouter())
+
+	r.Mount("/", routes.DummyRouter())
+
 	// if err := http.ListenAndServe(":8080", nil); err != nil {
 	//     log.Fatal(err)
 	// }
-	if err := http.ListenAndServeTLS(":8080", "/Users/noellim/customkeystore/server.cert", "/Users/noellim/customkeystore/server.key", nil); err != nil {
+	if err := http.ListenAndServeTLS(":8080", "/Users/noellim/customkeystore/server.cert", "/Users/noellim/customkeystore/server.key", r); err != nil {
 		log.Fatal(err)
 	}
