+++ test-server.go
@@ -1,0 +1,7 @@
+package main
+
+import "testing"
+
+func TestFunction1(t *testing.T) {
+
+}
